pyproject.toml
[tool.poetry]
name = "quantdevproject"
version = "0.1.0"
description = "A library for quantiative forecasting and trading"
authors = ["John Farrell <johndfarrell@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.1.2"
pandas = "^2.2.3"
joblib = "^1.4.2"
scikit-learn = "^1.5.2"
scipy = "^1.14.1"
statsmodels = "^0.14.4"
psycopg2 = "^2.9.9"
yfinance = "^0.2.44"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"
sec-edgar-api = "^1.1.0"
openmeteo-requests = "^1.3.0"
requests-cache = "^1.2.1"
retry-requests = "^2.0.0"
transformers = "^4.46.0"
torch = "^2.5.0"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
dash = "^2.18.2"
dash-bootstrap-components = "^1.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
